requirements.txt
absl-py==0.15.0
alembic==1.7.7
aliyun-log-python-sdk==0.7.8
appdirs==1.4.4
astunparse==1.6.3
bokeh==2.4.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
clang==5.0
coremltools==5.2.0
cycler==0.11.0
Cython==0.29.28
dateparser==1.1.1
dm-tree==0.1.7
elastic-transport==8.1.2
elasticsearch==8.1.2
fire==0.4.0
flatbuffers==1.12
fonttools==4.33.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h5py==3.1.0
idna==3.3
imagecorruptions==1.1.2
imageio==2.17.0
imgaug==0.4.0
importlib-metadata==4.11.3
Jinja2==3.1.1
jmespath==1.0.0
joblib==1.1.0
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras2onnx==1.7.0
kiwisolver==1.4.2
libclang==13.0.0
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
MNN==1.1.7
mpmath==1.2.1
networkx==2.8
numpy==1.19.5
oauthlib==3.2.0
onnx==1.11.0
onnxconverter-common==1.9.0
onnxruntime==1.10.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyparsing==3.0.8
PyQt5==5.14.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.7.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==5.3.1
qTimer==0.1.4
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==0.24.2
scipy==1.8.0
seaborn==0.11.2
Shapely==1.8.1.post1
six==1.15.0
SQLAlchemy==1.4.35
sympy==1.10.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.3
tensorflow-addons==0.16.1
tensorflow-cpu==2.6.3
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.3
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-model-optimization==0.5.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tf2onnx==1.9.3
tfcoreml==1.1
threadpoolctl==3.1.0
tidecv==1.0.1
tifffile==2022.4.22
tornado==6.1
tqdm==4.64.0
typeguard==2.13.3
typing-extensions==3.10.0.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.12.1
zipp==3.8.0
